net: ipa: Move GSI-specific functions to an ops struct
Add a gsi_ops struct and add all GSI-specific functions to it, then
use it to call those functions. This will make it easy to plug-in
a new BAM backend for use with IPA 2.* that will be added in the future.
In order to reuse gsi_channel_tre_max for the BAM backend, don't add it
to the ops struct but move its template to gsi_private.h for now.

This makes no functional change.

Signed-off-by: Yassine Oudjana <y.oudjana@protonmail.com>