net: ipa: gsi.h: Rename gsi to ipa_dma
Rename gsi.h to ipa_dma.h, and rename all of its contents currently
named gsi* that arent GSI-specific, and all of their occurences
everywhere to ipa_dma*. These structs and enums will be reused
for a new BAM backend that will be added later for IPA 2.* support
so it makes sense to give them more general names.

This makes no functional change.

Signed-off-by: Yassine Oudjana <y.oudjana@protonmail.com>

Here the new access functions of the gsi registers had to be used.

Signed-off-by: Bert Karwatzki <spasswolf@web.de>