arm64: dts: qcom: msm8953: add mpm node
This adds a devicetree node for the mpm interrupt controller
and a node for the sram needed by the controller. The interrupt mapping
data is taken from the fairphone-fp3 downstream kernel
(drivers/irqchip/qcom/mpm8953.c)
The controller is no yet usable as its powerdomain needs to be the parent
powerdomain of the CPU cluster. For this we first need to setup psci
poweromains as child nodes of the psci nodes.

Link: https://code.fairphone.com/projects/fairphone-3/gpl.html

Signed-off-by: Bert Karwatzki <spasswolf@web.de>