qrtr: smd: Enable waking up for calls.
This successfully wakes the fairphone-fp3 from s2idle when a call
is incoming. This is an adaption and extension of Caleb Connolly's patch
for glink. It recognizes incoming calls and messages by examining the
src_port of the qrtr protocol and sends a wakup event in these cases. I
don't no if the port number (which are reported by qrtr-lookup) are
universal so this is probably not portable, yet.

Link: https://lore.kernel.org/all/20230117142414.983946-1-caleb.connolly@linaro.org/
Link: https://gitlab.freedesktop.org/mobile-broadband/ModemManager/-/issues/694

Signed-off-by: Bert Karwatzki <spasswolf@web.de>